% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hey_siri}
\alias{hey_siri}
\title{A function that reports back today's date}
\usage{
hey_siri(name, month, day)
}
\arguments{
\item{name}{insert your preferred name}

\item{month}{the current month (chr.)}

\item{day}{the current day (num.)}
}
\description{
A function that reports back today's date
}
